feat: add topic creation and retrieval endpoints

+++ back/src/main/java/com/openclassrooms/mdd/service/UserService.java
@@ -16,7 +16,7 @@ public interface UserService {
     User authenticateUser(String email, String password) throws AuthenticationException;
 
     Optional<User> findUserByEmail(String email);
-    Optional<User> findUserById(final long id);
+    Optional<User> findUserById(final int id);
     String encodePassword(String password);
     void deleteUser(User user);
 }
